Tamper detection is now disabled and support for new conf.yml layout.
This change also fixes common problems with the extension installation. The problem was that parse_yaml could not read the file correctly due to problems with the conf.yml layout.

Identifier checks are also fixed now. The if statement that checks if the identifier is only lowercase and a-z was written incorrectly.