Finish the Installation instructions for README.md
Off-topic note: Blueprint was meant to be a paid add-on in it's early days. This means some parts of the commits may talk about "obfuscation for the license system". The license system has been scrapped since and is not planned to come back, we are trying to keep our code open. Working on the project was messy back then, and you may see some outside people make commits and me question what others were doing before this became a solo-project. Have fun laughing at the previous commits of us trying to figure out if this project is even possible to do.